AG-14318 Move eslint-rules/ into libraries/
