+++ libraries/ag-charts-eslint-rules/src/lint-change-detection-eslint-config.mjs
@@ -2,7 +2,7 @@
 import globals from 'globals';
 import tseslint from 'typescript-eslint';
 
-import changeDetection from '../../../eslint-rules/change-detection.mjs';
+import changeDetection from '../rules/change-detection.mjs';
 
 /** @type {import('eslint').Linter.FlatConfig[]} */
 export default [
