+++ libraries/ag-charts-eslint-rules/tsconfig.spec.json
@@ -3,6 +3,6 @@
   "compilerOptions": {
     "types": ["jest", "node"]
   },
-  "include": ["jest.config.ts", "src/**/*.test.ts", "src/**/*.spec.ts", "src/**/*.d.ts"],
-  "exclude": ["src/**/*.data.ts"]
+  "include": ["src/**/*.test.ts", "jest.config.ts", "src/**/*.spec.ts", "src/**/*.d.ts"],
+  "exclude": ["src/**/*.data.ts", "rules/*.mjs"]
 }
